requirements.txt
django
django_rest_framework
ipykernel
requests